Pretend to be the web player
Request the server with the same headers than the web player, in an
attempt to be as discreet as possible. I don't think it makes any
difference, but it was fun to implement.
Also, we query at the same period than the web player, just to be fair
with the server.

The goal of all these changes was to reduce the amount of timeouts and
connection errors on RadioNova. But the fact is that it happens also
with the web player, so the problem comes from their server, not from
our requests.